restructure files into directories, add tests

+++ src/schemas/InitialConfig.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel
-
-from pydantic import BaseModel, Field,ValidationError, Field
 from typing import Literal, List
+from pydantic import BaseModel, Field,ValidationError
+
 
 
 class InitialConfig(BaseModel):
@@ -16,11 +16,10 @@ class InitialConfig(BaseModel):
 	large_number_of_items_penalty_value: int
 	max_delivery_fee_value: int
 	free_delivery_eligible_cart_value: int
-	rush_hour_multiplier: int
+	rush_hour_multiplier: float
 	rush_hour_days: List[Literal[0, 1, 2, 3, 4, 5, 6]] = Field(min_length=1)
 	rush_hour_start_hour: int
 	rush_hour_end_hour: int
-
